Example code with dummy foils.
Example code on how to generate existential statements for each NP in an input sentence. Currently foils are manually handcrafted but we wish to update this script to use SpanBERT (or a similar model) to propose alternatives to NPs we wish to foil, a pretrained NLI model to ensure the foil does not entail the original existential statement, and GRUEN to make sure the foil is grammatical (grammaticality score > 0.7).